Moved emptying the store to the Integration Test Case.
Why?  Emptying the store is a healthy task that many different integration
tests will need access to in order to clean up after a test that adds configs
into the store.  Centralizing this process:

1. exposes it to all integration tests, ensuring that any test has it available for use.
2. keeps the code DRY
